# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "LocateThis API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/availability/": {
            "post": {
                "description": "Create a new availability with the provided begin and end times, and user ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "availabilities"
                ],
                "summary": "Create a new availability",
                "parameters": [
                    {
                        "description": "Availability information",
                        "name": "availability",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AvailabilityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AvailabilityResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/availability/availabilities": {
            "get": {
                "description": "Retrieve a list of all availabilities",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "availabilities"
                ],
                "summary": "Get all availabilities",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.AvailabilityResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/availability/user/{userID}": {
            "get": {
                "description": "Retrieve a list of availabilities associated with a specific user ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "availabilities"
                ],
                "summary": "Get availabilities by user ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.AvailabilityResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/availability/{id}": {
            "get": {
                "description": "Retrieve an availability by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "availabilities"
                ],
                "summary": "Get availability by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Availability ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AvailabilityResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an availability identified by its ID with the provided begin and end times, and user ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "availabilities"
                ],
                "summary": "Update an availability by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Availability ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Availability information",
                        "name": "availability",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AvailabilityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an availability identified by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "availabilities"
                ],
                "summary": "Delete an availability by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Availability ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.AvailabilityRequest": {
            "type": "object",
            "properties": {
                "date_begin": {
                    "type": "string"
                },
                "date_end": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.AvailabilityResponse": {
            "type": "object",
            "properties": {
                "date_begin": {
                    "type": "string"
                },
                "date_end": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}